+++ shared/API.h
@@ -238,7 +238,7 @@ public:
 	virtual int GetVehicleWheelColor(unsigned long guid) = 0; //Not implemented
 	virtual bool SetVehicleWheelType(unsigned long guid, int type) = 0; //Not implemented
 	virtual int GetVehicleWheelType(unsigned long guid) = 0; //Not implemented
-	virtual long GetVehicleDriver(unsigned long guid) = 0;
+	virtual int GetVehicleDriver(unsigned long guid) = 0;
 	virtual std::vector<unsigned int> GetVehiclePassengers(unsigned long guid) = 0;
 
 	virtual unsigned long CreateObject(long model, float x, float y, float z, float pitch, float yaw, float roll) = 0;
